// argefrontend/src/app/views/+basvuru/forms/model/arastirma-kapsam-model.ts
export class ArastirmaKapsam {
    amac?: string;
    genelBilgi?: string;
    arastirmaTuru?: string;
    arYapilacagiYer?: string;
    arEvrenOrneklem?: string;
    hipotez?: string;
    veriToplamaYontemi?: string;
    arastirmaKisitlari?: string;
    arastirmaTarihAraligi?: string;
    invazivmi?: string;
    invazAcikla?: string;
    tibbiDurumVarmi?: string;
    tibbiDurumAcikla?: string;
    arastirmaRuhsalSoru?: string;
    arastirmaRuhsalSoruAcikla?: string;
    gonulluKatilimSoru?: string;
    gonulluKatilimSoruAcikla?: string;
    bireyselYarar?: string;
    kurumsalYarar?: string;
    toplumsalYarar?: string;
    bilimselYarar?: string;
    destekleyicisiVarmi?: string;
    destekleyicisiAcikla?: string;
    bizdenTalepVarmi?: string;
    bizdenTalepAcikla?: string;
}